Minor changes
Generic
 * Menu -- Tools -- 'Create Record Session - Template file' Wizard...' now writes a template file with comments (and all default values are commented out). You can choose to sort the CM Entries by name or have it in creation order!
 * PCS Writers now also Writes "Target Info: <where did we store the data>" in the log file (for tracability)
 * AppDir got some improvements for the upcomming Windows: DbxInstaller
 * MailUtil: If no 'to' address has been given, fallback on property '{cfgName|mail}.to'

DbxCentral
 * Counter Details -- Remember current tab name within a "tab group" (move more fluently between tabs, and remember the last visited CM)

SqlServerTune
 * Daily Summary Report -- SqlServerJobScheduler -- Added column 'scheduledLast24hChart', which is a "gantt" that tries to visualize when a job was executed during the day.

+++ src/com/dbxtune/cm/sqlserver/CmPerfCounters.java
@@ -3085,9 +3085,9 @@ extends CountersModel
 		CmSettingsHelper.Type isAlarmSwitch = CmSettingsHelper.Type.IS_ALARM_SWITCH;
 		
 		list.add(new CmSettingsHelper("FreeListStalls",            isAlarmSwitch, PROPKEY_alarm_FreeListStalls      , Double .class, conf.getDoubleProperty(PROPKEY_alarm_FreeListStalls      , DEFAULT_alarm_FreeListStalls      ), DEFAULT_alarm_FreeListStalls      , "If 'Free list stalls/sec' is greater than 0.5 then send 'AlarmEventPerfCounterWarning'." ));
-		list.add(new CmSettingsHelper("FreeListStalls RaiseDelay",                PROPKEY_alarm_FreeListStalls_delay, Integer.class, conf.getDoubleProperty(PROPKEY_alarm_FreeListStalls_delay, DEFAULT_alarm_FreeListStalls_delay), DEFAULT_alarm_FreeListStalls_delay, "If 'Free list stalls/sec' is true and has been so for 2 minutes then proceed with the Alarm Raise." ));
+		list.add(new CmSettingsHelper("FreeListStalls RaiseDelay",                PROPKEY_alarm_FreeListStalls_delay, Integer.class, conf.getIntProperty   (PROPKEY_alarm_FreeListStalls_delay, DEFAULT_alarm_FreeListStalls_delay), DEFAULT_alarm_FreeListStalls_delay, "If 'Free list stalls/sec' is true and has been so for 2 minutes then proceed with the Alarm Raise." ));
 		list.add(new CmSettingsHelper("LazyWrites",                isAlarmSwitch, PROPKEY_alarm_LazyWrites          , Double.class , conf.getDoubleProperty(PROPKEY_alarm_LazyWrites          , DEFAULT_alarm_LazyWrites          ), DEFAULT_alarm_LazyWrites          , "If 'Lazy writes/sec' is greater than 20 then send 'AlarmEventPerfCounterWarning'." ));
-		list.add(new CmSettingsHelper("LazyWrites RaiseDelay",                    PROPKEY_alarm_LazyWrites_delay    , Integer.class, conf.getDoubleProperty(PROPKEY_alarm_LazyWrites_delay    , DEFAULT_alarm_LazyWrites_delay    ), DEFAULT_alarm_LazyWrites_delay    , "If 'Lazy writes/sec' is true and has been so for 3 minutes then proceed with the Alarm Raise." ));
+		list.add(new CmSettingsHelper("LazyWrites RaiseDelay",                    PROPKEY_alarm_LazyWrites_delay    , Integer.class, conf.getIntProperty   (PROPKEY_alarm_LazyWrites_delay    , DEFAULT_alarm_LazyWrites_delay    ), DEFAULT_alarm_LazyWrites_delay    , "If 'Lazy writes/sec' is true and has been so for 3 minutes then proceed with the Alarm Raise." ));
 
 		return list;
 	}
